Add load balancer components class
